+++ lib/braille.txt
@@ -1,0 +1,3 @@
+0.0.0.0.0....0..000....00....00.0..00000
+00.00.0..0..0...00.0..00.0..00..000..000
+....0.0.0.........0...0.0...0.000...0...
