+++ lib/message.txt
@@ -1,1 +1,1 @@
-hello world
+hello i go to turing
